Avoid overflow in HRES mode3 scheduling
Change hres_per_node from uint16_t to uint32_t to avoid overflow when a
value larger than 65535 is configured or requested. This affects mode3 HRES.

+++ src/slurmctld/licenses.h
@@ -67,7 +67,7 @@ typedef struct {
 	uint32_t *avail_hres;
 	uint32_t *avail_hres_orig;
 	uint16_t depth;
-	uint16_t hres_per_node;
+	uint32_t hres_per_node;
 	uint16_t layers_cnt; /* size of avail_hres */
 	hres_leaf_t *leaf;
 	uint16_t leaf_cnt; /* size of leaf */
